#126 refactor: 재사용하지 않는 LivingBtn 컴포넌트를 LivingInfo 파일 내에 선언하여 활용
